Refactor: break data fetch into hook
Refactor: break data fetch into hook for dynamic GeoJSON data

Separates fetch function from MyMapboxPopupDynamicLayers into own hook to be called from MyMapboxPopupDynamicLayers.
